Refactoring in preparation for new toolbar implementation
