+++ packages/js-draw/src/Editor.ts
@@ -7,7 +7,7 @@ import Viewport from './Viewport';
 import EventDispatcher from './EventDispatcher';
 import { Point2, Vec2 } from './math/Vec2';
 import Vec3 from './math/Vec3';
-import HTMLToolbar from './toolbar/HTMLToolbar';
+import DropdownToolbar from './toolbar/DropdownToolbar';
 import { RenderablePathSpec } from './rendering/renderers/AbstractRenderer';
 import Display, { RenderingMode } from './rendering/Display';
 import SVGRenderer from './rendering/renderers/SVGRenderer';
@@ -31,6 +31,7 @@ import BackgroundComponent, { BackgroundType } from './components/BackgroundComp
 import sendPenEvent from './testing/sendPenEvent';
 import KeyboardShortcutManager from './shortcuts/KeyboardShortcutManager';
 import KeyBinding from './shortcuts/KeyBinding';
+import AbstractToolbar from './toolbar/AbstractToolbar';
 
 export interface EditorSettings {
 	/** Defaults to `RenderingMode.CanvasRenderer` */
@@ -327,8 +328,8 @@ export class Editor {
 	 * Creates a toolbar. If `defaultLayout` is true, default buttons are used.
 	 * @returns a reference to the toolbar.
 	 */
-	public addToolbar(defaultLayout: boolean = true): HTMLToolbar {
-		const toolbar = new HTMLToolbar(this, this.container, this.localization);
+	public addToolbar(defaultLayout: boolean = true): AbstractToolbar {
+		const toolbar = new DropdownToolbar(this, this.container, this.localization);
 
 		if (defaultLayout) {
 			toolbar.addDefaults();
